validate xroute enums
